Delete header files with C standard style names
Delete all files with names of C standard library style and create
utility headers under include/urd. All #include directive in header
files are removed so that the compiler can work more efficiently with
the cost of requiring source files to order their include correctly.

+++ kernel/include/urd/syscall.h
@@ -4,7 +4,7 @@
 #ifndef URD_KERNEL_SYSCALL_H_
 #define URD_KERNEL_SYSCALL_H_
 
-long long int syscall(int number, ...);
+int syscall(int number, ...);
 
 int syscall_init(void);
